Relax mem ordering more when exiting a crit. sect.
When exiting a critical section, `in_critical`, which tracks the number
of reentrants, is decremented.  Originally, the memory ordering of the
decrement was Release if the decremented value is bigger than 1, but
this commit changes the threshold to 0.  This is because the only reason
for the Release ordering is for `try_collect` to see the change, but it
only distinguishes nonzero values from the zero.