Fix two bugs that prevent memory reclamation.
First, consider entries in the "new" bag when determining the size of
the garbage. Without this, no collection will ever be triggered, since
the threshold will never be reached (as new garbage is added to the
"new" bag).

Second, do a local collection when try_collect succeeds. Without this,
if one thread is usually or always the thread triggering collection
(such as in a single-consumer scenario), that thread will never
collect its garbage.