Fix a double-free bug in lazy freeing of Participants.
Add an `unlinked` field to Participant in order to ensure that only
one thread frees them while iterating over Participants.

A double free could occur if some node `A` is disconnected while a
thread is looking at it (because it became inactive) and the node that
followed it inthe list is also inactive. Then, if the node following
`A` is also inactive, both the thread looking at `A` and the thread
that just disconnected `A` will attempt to disconnect `A`'s successor,
and both may succeed.

To prevent multiple threads from freeing a node in this case, add an
`unlinked` field that we CAS against to ensure we only free the node
once.