Fix multiple segfault-causing bugs
This commit closes two major holes, one related to the use of local
garbage bags, and the other to reclaiming participant data from exited
threads. After making these changes, I am no longer able to reproduce
segfaults or assertion failures. (I had previously build some aggressive
testing that was able to consistently segfault in a few seconds on every try.)

The use of local bags was creating a problem in which garbage could be
collected too early. It's crucial that when `unlinked` is called, the
data to be reclaimed is associated with the *current global
epoch* (and **not** the current local snapshot of the epoch). The reason
is that other threads may have outstanding pointers to the data, and may
be an epoch ahead; when the thread calling `unlinked` unpins, the epoch
could move forward again, causing the `unlinked` data to be collected
while pointers to it remain.

I was aware of this issue with the epoch algorithm, but Crossbeam
attempted to optimize things by using local bags of garbage nodes. These
local bags, however, did not manage to avoid the above problem.

This commit reverts the use of local bags (which has caused other
problems), returning to something closer to the original epoch algorithm
for the time being.

There was also a bug in freeing participant data from threads that have
exited after being involved in epoch management -- essentially, the code
removing a node from the middle of a linked-list was subject to race
conditions in which a previously-`unlinked` node could suddenly
reappear.

For the time being, I have simply removed all reclamation of participant
nodes. This is a memory leak, but a rather small one: it's proportional
to the number of threads that have ever used epochs and have since
terminated.

In a follow-up PR, I will try to roll out a working algorithm for
participant reclamation, likely based on marked pointers.