Implement the `Debug` trait for all public types (#65)
This commit implements `std::fmt::Debug` for all public types. In most cases,
this was simply adding `#[derive(Debug)]` to the type and the types of its
members. However, there are two exceptions, where `Debug` is implemented
manually because it cannot be derived: (1) `Scope<'a>`, which uses `FnBox` which
is not `Debug`, and (2) `ScopedJoinHandle<T>`, which uses `JoinState` which uses
`std::thread::JoinHandle<T>` which is not `Debug`.

Fixes #62.