Added MsQueue::push_batch
push_batch pushes a contiguous group of items onto the queue. This slightly changes the invariant, as a blocked queue can now have a batch node at the head.
This doesn't change how pop works, but it does alter how push works, as push now waits on the push_batch to complete.