Fixes ordering bug in ms queue (#98)
`Signal.ready` is being used to guard access to the data slot. This
requires using acquire / release ordering vs. relaxed.

I believe that the original code was written with the assumption that
`thread::park() / unpark()` was able to provide the necessary ordering,
but this is not true given that park() can spuriously wakeup, which
means that there is a chance (although, a small one) for the memory to
not be in a consistent state.

Fixes #97