Introduce owned epoch guards (`Pinned<T>`).
Currently every epoch guard is usually used through a reference, making
it somewhat inconvinient for certain usecases, especially when you need
to expose an API, where you don't want the downstream user to have to
interact with the crossbeam API.

This patch solves this issue by introducing the `Pinned<T>` type, which
wraps `T` in a manner that only allow access through immutable
reference. `Pinned<T>` also holds an `epoch::Guard` meaning that `T` is
essentially "bound" to this guard.

Initially, I tried my hands at a `pin_then` method taking a closure, but
it proved inconvinient and possibly unsound, so instead I let
`Pinned<T>` be created through the methods of `Owned<T>` and
`Shared<T>`.

TLDR. This gets rid of the lifetime hell by providing a wrapper type,
which guarantees that a `Guard` exists while the inner value is held.