Let ZerosValid be a regular non-auto trait, also in nightly
This bug is related to the crate feature "nightly" only.

It's an unintended bug, but ZerosValid was defined for every type and
did not serve its purpose. For example, `Vec<i32>` would implement
ZerosValid when it was an auto trait, and that was a bug -- all-zeros is
not a correct way to create a Vec; its pointer is always nonnull.

It does not seem to be possible to exclude all types by default and then
one by one whitelist types that are valid.

We could attempt a blacklist:

  impl<T: ?Sized> !ZerosValid for *const T { }
  impl<T: ?Sized> !ZerosValid for *mut T { }
  impl<'a, T: ?Sized> !ZerosValid for &'a T { }
  impl<'a, T: ?Sized> !ZerosValid for &'a mut T { }

And this excludes a lot of problematic types, like Vec, Box, Mutex etc!
But it is not a complete solution.

CachePadded::zeroed allows creating values from nothing as long as they
are ZerosValid. Let's say for example `struct UnforgeableToken(u64);`

With CachePadded::zeroed we could then create
`CachePadded<UnforgeableToken>` and use DerefMut to swap one of our valid
tokens with a forged, all-zeros one.