Add a sharded reader-writer lock (#48)
This PR introduces `ShardedLock`, which is a variant of `RwLock` where concurrent read operations don't contend as much.

`ShardedLock` consists of a bunch of smaller `RwLock`s called *shards*.

A read operation read-locks a single shard only. The shard is chosen based on the current thread so that contention is reduced as much as possible.

A write operation has to write-lock every shard in order. That means we trade faster reads for slower writes.

Another way of looking at it is: sharded locking is just poor man's hardware lock elision. While `parking_lot`'s `RwLock` does use HLE on platforms that support it, we have to resort to `ShardedLock` on others.

This PR is basically just a port of [another one](https://github.com/tokio-rs/tokio/pull/517) that added a sharded RW lock to `tokio-reactor` in order to improve the performance of contended reads. I was told that `ShardedLock` would also be useful in [salsa](https://github.com/nikomatsakis/salsa). Let's add it to Crossbeam so that anyone can use it.