crossbeam-epoch: Add Shared::deref_mut()
It is possible that the user would like to have one thread to be able to
update the value of the object pointed by Atomic. For that, they will
need to have a mutable refernce towards that object.

This is an unsafe function because there is no guarantee that there are
more threads attempting to read/write from/to the actual object in the
same time.
In order to proivde safety the user must be able to know that there are
no concurrent accesses towards the object itself and/or use
crossbeam_utils::atomic::AtomicCell to provide the needed guarantees.