channel: implementation for receiving multiple items at once
This commit implements `try_recv_many`, which allows a caller to receive
multiple items in a single call.  It is provided generally for all
channels, but has a specific implementation override for bounded
channels, and a few caveats:

- there are no timeout-based or blocking variants
- this will do almost nothing for some channels like after/zero because
it uses try semantics so it will give up quickly waiting for more items
- callers must allocate an `Extend` capable container themselves
- mixed recv/recv_many performance may be suboptimal

This method fills a niche use case: single consumers that can benefit
from grabbing and processing multiple items at a time.

Signed-off-by: Toby Lawrence <toby@nuclearfurnace.com>