SeqLock: Aborted writes should not update the stamp.
This is minor, but I noticed it on a read of the code. `abort()` would
write back the old stamp, but would then immediately drop `self`,
which would bump the stamp anyways.

I included a test; SeqLock isn't exported, so it wasn't possible to
add this to the integratio suite so I added it directly in the file.