Don't call assume_init on Deferred's Data
In the crossbeam-deque test suite, a Deferred is created from a FnOnce
which is smaller than the Data. This makes the call to assume_init()
immediate UB (the reference to it created upon call is probably UB too).
Since Data doesn't have a Drop impl, we can just leave it in the MaybeUninit
wrapper.