When building for linux, replace all vendors with 'unknown'
Some build systems such as the Yocto project define custom
rust targets, of the form 'arch-somevendor-linux-gnu/musl'
(only the 'somevendor' part is custom, the rest matches rust's
definitions).

This causes mismatches with crossbeam's lists of targets
which are built from the standard rust target list, and
always have 'unknown' as the vendor.

This change simply replaces such custom vendors with 'unknown' when
the third component of the target is 'linux'.